Temp version, that may be final.
Could not manage to link __device__ functions in raycast_common.cu to
program in raycast_gpu.cu. The nvcc flag "-relocatable-device-code
true" enabled this, but is only available to nvcc 5.0+, and not 4.0
which the school machines run.

For now, I'm just duplicating code for each program instead of doing it
"the smart way".