Added functionality -> time_stack,
so that I can easily store kernel calls and later count them/sum their
total time. A todo for later would be to write a sum function for the
stack, which sums in the right order for floats to be accurate.

Also cleaned up code some, removing comments and commented lines of
code.